+++ src/pymyhondaplus/__init__.py
@@ -1,6 +1,6 @@
 """Unofficial Honda Connect Europe (My Honda+) API client library."""
 
-from .api import AuthTokens, CommandResult, EVStatus, HondaAPI, HondaAPIError, HondaAuthError, SubscriptionService, Subscription, UIConfiguration, UserProfile, Vehicle, VehicleCapabilities, compute_trip_stats, parse_charge_schedule, parse_climate_schedule, parse_ev_status
+from .api import AuthTokens, CommandResult, EVStatus, Geofence, HondaAPI, HondaAPIError, HondaAuthError, SubscriptionService, Subscription, UIConfiguration, UserProfile, Vehicle, VehicleCapabilities, compute_trip_stats, parse_charge_schedule, parse_climate_schedule, parse_ev_status
 from .auth import DeviceKey, HondaAuth, encrypt_request
 from .storage import SecretStorage, get_storage
 
@@ -8,6 +8,7 @@ __all__ = [
     "AuthTokens",
     "CommandResult",
     "EVStatus",
+    "Geofence",
     "Subscription",
     "SubscriptionService",
     "UIConfiguration",